refactor: split rich editor code
