+++ src/features/NoteEditor/RichEditor/RichEditor.tsx
@@ -15,7 +15,7 @@ import { FormattingNode } from './nodes/FormattingNode';
 import { ImageNode } from './nodes/ImageNode';
 import { RawNode } from './nodes/RawNode';
 import theme from './PlaygroundEditorTheme';
-import { RichEditorContent, RichEditorContentProps } from './RichEditorContentProps';
+import { RichEditorContent, RichEditorContentProps } from './RichEditorContent';
 
 // TODO: move lines with alt+arrows
 // TODO: support videos
